feat: excluindo usuarios

+++ backend/src/services/users/DeleteUserService.ts
@@ -9,7 +9,10 @@ export const DeleteUserService = async ({ userId }: { userId: string }) => {
         const user = await Prisma.user.delete({
             where: { id: userId }
         })
-        await unlink(resolve(__dirname, '..', '..', '..', 'uploads', user.avatar))
+
+        if (user.avatar !== null) {
+            await unlink(resolve(__dirname, '..', '..', '..', 'uploads', user.avatar))
+        }
         return user
     } catch (error) {
         console.log(error)
